pyproject.toml
[project]
name = "mdq"
version = "0.1.0"
description = "Market data quant research: intraday level reactions on SPY / QQQ"
authors = [{ name = "Francisco Castillo" }]
requires-python = ">=3.12,<3.13"
dependencies = [
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "pandas>=2.2",
    "pyarrow>=16.0",
    "numpy>=1.26",
    "pytz>=2024.1",
    "tqdm>=4.66",
    "alpaca-py (>=0.30)",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.poetry]
packages = [{ include = "mdq", from = "src" }]

[build-system]
requires = ["poetry-core>=2.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
